+++ test/test_scripts/bash_aliases_argmap_test.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+# TODO: Rename file as bash_test_functions.lib.sh
+
 # For running casual tests and checks.
 # These aliases are not considered part of a public API, and therefore updates may change them without warning.
 
@@ -24,10 +26,10 @@ alias argmh2='rm $PATH_OUTPUT_LOCAL/html/example1-clearly-false-white-swan-simpl
 alias argmhmeta='rm $PATH_OUTPUT_LOCAL/html/example1-clearly-false-white-swan-simplified-meta-mapjs.html; 2hf $INPUT_FILE_MD_META'
 alias argmp='rm $PATH_OUTPUT_LOCAL/example1-clearly-false-white-swan-simplified.pdf; md2pdf $INPUT_FILE_MD'
 # TODO: simplify argmph call
-alias argmph='rm $PATH_OUTPUT_LOCAL/example.pdf; rm $PATH_OUTPUT_LOCAL/header.tex; $PATH_DIR_ARGMAP_LUA/argmap2tikz.lua -i > $PATH_OUTPUT_LOCAL/header.tex; pandoc $INPUT_FILE_MD -o $PATH_OUTPUT_LOCAL/example.pdf --lua-filter pandoc-argmap.lua --pdf-engine lualatex --include-in-header $PATH_OUTPUT_LOCAL/header.tex --data-dir=$CONDA_PREFIX/share/pandoc; echo "Generated: $PATH_OUTPUT_LOCAL/example.pdf"'
-alias argmf='rm $PATH_OUTPUT_LOCAL/html/example1-clearly-false-white-swan-simplified-1mapjs-fragment.html; md2htm $WORKSPACE/test/input/markdown/example1-clearly-false-white-swan-simplified-1mapjs.md'
-alias argt='$WORKSPACE/test/test_scripts/tests.sh'
-alias argth='$WORKSPACE/test/test_scripts/tests.sh html'
+alias argmph='rm $PATH_OUTPUT_LOCAL/example.pdf; rm $PATH_OUTPUT_LOCAL/header.tex; $PATH_LUA_ARGMAP/argmap2tikz.lua -i > $PATH_OUTPUT_LOCAL/header.tex; pandoc $INPUT_FILE_MD -o $PATH_OUTPUT_LOCAL/example.pdf --lua-filter pandoc-argmap.lua --pdf-engine lualatex --include-in-header $PATH_OUTPUT_LOCAL/header.tex --data-dir=$PATH_ENVIRONMENT_GLOBAL/share/pandoc; echo "Generated: $PATH_OUTPUT_LOCAL/example.pdf"'
+alias argmf='rm $PATH_OUTPUT_LOCAL/html/example1-clearly-false-white-swan-simplified-1mapjs-fragment.html; md2htm $PATH_ARGMAP_ROOT/test/input/markdown/example1-clearly-false-white-swan-simplified-1mapjs.md'
+alias argt='$PATH_ARGMAP_ROOT/test/test_scripts/tests.sh'
+alias argth='$PATH_ARGMAP_ROOT/test/test_scripts/tests.sh html'
 
 __init_tests() {
   COLOUR='true'
@@ -53,6 +55,8 @@ __init_tests() {
     COL_FAIL='\033[0;31m' # Red
     COL_RESET='\033[0m'   # No Color
   fi
+
+  mkdir --parent "$(dirname "$(getvar PATH_TEST_LOG)")"
 }
 
 check_var_value() {
@@ -68,7 +72,7 @@ __test() {
   if "$@" >>"$(getvar PATH_TEST_LOG)"; then
     echo -e "${COL_PASS}Pass${COL_RESET}"
   else
-    echo -e "$PRE ${COL_FAIL}Fail${COL_RESET}"
+    echo -e "$PRE ${COL_FAIL}Fail${COL_RESET}" >&2
     FAILCOUNT=$((FAILCOUNT + 1))
   fi
 
@@ -84,34 +88,47 @@ test_function() {
 test_getvar() {
   __init_tests
   func=getvar
-  test_function "$func" bla "$([[ $? -eq 1 ]])"                                                                                          #1
-  test_function "$func" bla.bla "$([[ $? -eq 1 ]])"                                                                                      #2
-  test_function "$func" DIR_CONFIG config                                                                                                #3
-  test_function "$func" PATH_FILE_CONFIG_ARGMAP_PROCESSED /home/s6mike/git_projects/argmap/config/processed/config-argmap-processed.yaml #4
-  test_function "$func" PORT_DEV_SERVER 9001                                                                                             #5
-  test_function "$func" PATH_DIR_MAPJS_ROOT /home/s6mike/git_projects/argmap/mapjs                                                       #6
-  test_function "$func" DEFAULT_LANGUAGE en
-  test_function "$func" DIR_OUTPUT output
-  test_function "$func" PATH_OUTPUT_LOCAL /home/s6mike/git_projects/argmap/test/output
-  test_function "$func" DIR_MAPJS_JSON mapjs-json #7
-  test_function "$func" PATH_DIR_PUBLIC_MAPJS_JSON /home/s6mike/git_projects/argmap/test/output/mapjs-json
-  test_function "$func" node "class: mapjs-node" #8
-  test_function "$func" node.class mapjs-node    #9
+  set +o nounset
+  test_function "$func" bla "$([[ $? -eq 1 ]])"                                                         #1
+  test_function "$func" bla.bla "$([[ $? -eq 1 ]])"                                                     #2
+  test_function "$func" DIR_CONFIG config                                                               #3
+  test_function "$func" PATH_FILE_CONFIG_ARGMAP_PROCESSED config/processed/config-argmap-processed.yaml #4
+  test_function "$func" PORT_DEV_SERVER 9001                                                            #5
+  test_function "$func" PATH_DIR_MAPJS_ROOT mapjs                                                       #6
+  test_function "$func" DEFAULT_LANGUAGE en                                                             #7
+  test_function "$func" DIR_OUTPUT output                                                               #8
+  test_function "$func" PATH_OUTPUT_LOCAL test/output                                                   #9
+  test_function "$func" DIR_MAPJS_JSON mapjs-json                                                       #10
+  test_function "$func" PATH_MAPJS_JSON_LOCAL test/output/mapjs-json                                    #11
+  test_function "$func" node "class: mapjs-node"                                                        #12
+  test_function "$func" node.class mapjs-node                                                           #13
   remember=$LIST_FILES_CONFIG_INPUT
   unset LIST_FILES_CONFIG_INPUT
   test_function "$func" LIST_FILES_CONFIG_INPUT \
-    "/home/s6mike/git_projects/argmap/config/environment-argmap.yaml /home/s6mike/git_projects/argmap/config/config-argmap-paths.yaml /home/s6mike/git_projects/argmap/config/config-argmap.yaml /home/s6mike/git_projects/argmap/config/PRIVATE-environment-argmap.yaml /home/s6mike/git_projects/argmap/mapjs/config/environment-mapjs.yaml /home/s6mike/git_projects/argmap/mapjs/config/config-mapjs-paths.yaml /home/s6mike/git_projects/argmap/mapjs/config/config-mapjs.yaml" \
-    "-l" #10-14
+    $'config/environment-argmap.yaml\nconfig/config-argmap-paths.yaml\nconfig/config-argmap.yaml\nmapjs/config/environment-mapjs.yaml\nmapjs/config/config-mapjs-paths.yaml\nmapjs/config/config-mapjs.yaml' #10-14 -l is (de)list mode for __getvar_from_yaml(), but it doesn't seem to work now
   LIST_FILES_CONFIG_INPUT=$remember
+  set -o nounset
 }
 
 test_get_site_path() {
   __init_tests
   func=__get_site_path
+  set +o nounset
   test_function "$func" test/input/markdown/2-maps-swan-donkey.md input/markdown/2-maps-swan-donkey.md
   test_function "$func" /home/s6mike/git_projects/argmap/test/input/markdown/2-maps-swan-donkey.md input/markdown/2-maps-swan-donkey.md
   test_function "$func" test/output/html/example1-clearly-false-white-swan-simplified-1mapjs-fragment.html output/html/example1-clearly-false-white-swan-simplified-1mapjs-fragment.html
+  set -o nounset
+}
+
+test_make_mapjs_dependencies() {
+  file_html="$1"
+  file_json="$2"
+  html_open="$3"
+  make HTML_OPEN="false" "$file_html"
+  rm "$file_json"
+  make HTML_OPEN="$html_open" "$file_html"
+  ls "$file_json"
 }
 
 export -f __init_tests check_var_value __test
-export -f test_function test_getvar test_get_site_path
+export -f test_function test_getvar test_get_site_path test_make_mapjs_dependencies